Moved main method to DependencyTreeFormatterMain
